Refactor mappings
