+++ _gulp/helpers.js
@@ -1,4 +1,6 @@
 const gutil = require('gulp-util');
+const path = require('path');
+const shell = require('shelljs');
 
 const throwError = (plugin, error) => {
   throw new gutil.PluginError({
@@ -7,6 +9,18 @@ const throwError = (plugin, error) => {
   }, { showStack: true });
 }
 
+const touchFile = (filePath) => {
+  const parsed = path.parse(filePath);
+  if (!shell.test('-d', parsed.dir)) {
+    shell.mkdir('-p', parsed.dir);
+  }
+  if (!shell.test('-f', filePath)) {
+    shell.touch(filePath);
+  }
+}
+
+
 module.exports = {
-  gulpErr: throwError
+  gulpErr: throwError,
+  touch: touchFile
 }
